Replaced favicon.
Replaced old favicon to remove any association with Eric S. Raymond.
New Favicon is still based on the idea of using a Glider, but is
fully custom.  The dots are all different colors, like different types of people.
And it kinda looks like an H for hacking! or like a presentation, or even
a person sitting at a computer.  Also, a glider in this configuration,
moves up and to the right.  All improvements over the initial design.