Moves lambda source code hash to private variable
Deploying lambdas should not require making a new commit to this repository to update
the source code hash.  However, we do desire to track the source code hash, in order to keep a record
of the versions and times we deployed to production, and check for accidental drift.
So we have moved this value out of the repository, and into our private configuration management tool.
This decouples our deployment history from this repository.